+++ passkit/lib/src/pkpass/barcode.dart
@@ -3,7 +3,7 @@ import 'package:json_annotation/json_annotation.dart';
 part 'barcode.g.dart';
 
 /// Information about a pass’s barcode.
-@JsonSerializable()
+@JsonSerializable(includeIfNull: false)
 class Barcode {
   Barcode({
     this.altText,
@@ -18,15 +18,18 @@ class Barcode {
 
   /// Optional. Text displayed near the barcode. For example, a human-readable
   /// version of the barcode data in case the barcode doesn’t scan.
+  @JsonKey(name: 'altText')
   final String? altText;
 
   /// Required. Barcode format. For the barcode dictionary, you can use only the
   /// following values: PKBarcodeFormatQR, PKBarcodeFormatPDF417, or
   /// PKBarcodeFormatAztec. For dictionaries in the barcodes array, you may also
   /// use PKBarcodeFormatCode128.
+  @JsonKey(name: 'format')
   final PkPassBarcodeType format;
 
   /// Required. Message or payload to be displayed as a barcode.
+  @JsonKey(name: 'message')
   final String message;
 
   /// Required. Text encoding that is used to convert the message from the
@@ -34,6 +37,7 @@ class Barcode {
   /// The value is typically iso-8859-1, but you may use another encoding that
   /// is supported by your barcode scanning infrastructure.
   // IANA character set name, as a string
+  @JsonKey(name: 'messageEncoding')
   final String messageEncoding;
 
   /// Converts this instance to a JSON object