+++ passkit/lib/src/pkpass/location.dart
@@ -3,7 +3,7 @@ import 'package:json_annotation/json_annotation.dart';
 part 'location.g.dart';
 
 /// Information about a location.
-@JsonSerializable()
+@JsonSerializable(includeIfNull: false)
 class Location {
   Location({
     this.altitude,
@@ -18,16 +18,20 @@ class Location {
   Map<String, dynamic> toJson() => _$LocationToJson(this);
 
   /// Optional. Altitude, in meters, of the location.
+  @JsonKey(name: 'altitude')
   final double? altitude;
 
   /// Required. Latitude, in degrees, of the location.
+  @JsonKey(name: 'latitude')
   final double latitude;
 
   /// Required. Longitude, in degrees, of the location.
+  @JsonKey(name: 'longitude')
   final double longitude;
 
   /// Optional. Text displayed on the lock screen when the pass is currently
   /// relevant. For example, a description of the nearby location such as
   /// “Store nearby on 1st and Main.”
+  @JsonKey(name: 'relevantText')
   final String? relevantText;
 }